+++ src/Yasumi/data/translations/allSaintsDay.php
@@ -18,5 +18,6 @@ return [
     'it_IT' => 'Festa di Tutti i Santi',
     'fr_FR' => 'La Toussaint',
     'es_ES' => 'Día de todos los Santos',
-    'sv_SE' => 'alla helgons dag'
+    'sv_SE' => 'alla helgons dag',
+    'fi_FI' => 'Pyhäinpäivä'
 ];