+++ src/Yasumi/data/translations/ascensionDay.php
@@ -18,5 +18,6 @@ return [
     'fr_FR' => 'Ascension',
     'da_DK' => 'Kristi Himmelfartsdag',
     'nb_NO' => 'Kristi himmelfartsdag',
-    'sv_SE' => 'Kristi himmelsfärds dag'
+    'sv_SE' => 'Kristi himmelsfärds dag',
+    'fi_FI' => 'Helatorstai'
 ];