+++ src/Yasumi/data/translations/christmasDay.php
@@ -20,5 +20,6 @@ return [
     'es_ES' => 'Navidad',
     'da_DK' => '1. Juledag',
     'nb_NO' => 'Første juledag',
-    'sv_SE' => 'juldagen'
+    'sv_SE' => 'juldagen',
+    'fi_FI' => 'Joulupäivä'
 ];