+++ src/Yasumi/data/translations/easterMonday.php
@@ -20,5 +20,6 @@ return [
     'es_ES' => 'Lunes de Pascua',
     'da_DK' => '2. Påskedag',
     'nb_NO' => 'Andre påskedag',
-    'sv_SE' => 'annandag påsk'
+    'sv_SE' => 'annandag påsk',
+    'fi_FI' => '2. pääsiäispäivä'
 ];