+++ src/Yasumi/data/translations/secondChristmasDay.php
@@ -17,5 +17,7 @@ return [
     'nl_BE' => 'Tweede Kerstdag',
     'da_DK' => '2. Juledag',
     'nb_NO' => 'Andre juledag',
-    'sv_SE' => 'annandag jul'
+    'sv_SE' => 'annandag jul',
+    'fi_FI' => '2. joulupäivä'
+
 ];