+++ tests/Finland/EasterTest.php
@@ -1,0 +1,46 @@
+<?php
+/**
+ *  This file is part of the Yasumi package.
+ *
+ *  Copyright (c) 2015 - 2016 AzuyaLabs
+ *
+ *  For the full copyright and license information, please view the LICENSE
+ *  file that was distributed with this source code.
+ *
+ *  @author Sacha Telgenhof <stelgenhof@gmail.com>
+ */
+
+namespace Yasumi\Tests\Finland;
+
+use DateTime;
+use DateTimeZone;
+
+/**
+ * Class containing tests for Easter in Finland.
+ */
+class EasterTest extends FinlandBaseTestCase
+{
+    /**
+     * The name of the holiday to be tested
+     */
+    const HOLIDAY = 'easter';
+
+    /**
+     * Tests the holiday defined in this test.
+     */
+    public function testHoliday()
+    {
+        $year = 1677;
+        $this->assertHoliday(self::REGION, self::HOLIDAY, $year,
+            new DateTime("$year-4-15", new DateTimeZone(self::TIMEZONE)));
+    }
+
+    /**
+     * Tests translated name of the holiday defined in this test.
+     */
+    public function testTranslation()
+    {
+        $this->assertTranslatedHolidayName(self::REGION, self::HOLIDAY, $this->generateRandomYear(),
+            [self::LOCALE => 'Pääsiäispäivä']);
+    }
+}