+++ tests/Finland/FinlandBaseTestCase.php
@@ -1,0 +1,40 @@
+<?php
+/**
+ *  This file is part of the Yasumi package.
+ *
+ *  Copyright (c) 2015 - 2016 AzuyaLabs
+ *
+ *  For the full copyright and license information, please view the LICENSE
+ *  file that was distributed with this source code.
+ *
+ *  @author Sacha Telgenhof <stelgenhof@gmail.com>
+ */
+
+namespace Yasumi\Tests\Finland;
+
+use PHPUnit_Framework_TestCase;
+use Yasumi\Tests\YasumiBase;
+
+/**
+ * Base class for test cases of the Finland holiday provider.
+ */
+abstract class FinlandBaseTestCase extends PHPUnit_Framework_TestCase
+{
+    use YasumiBase;
+
+    /**
+     * Name of the region (e.g. country / state) to be tested
+     */
+    const REGION = 'Finland';
+
+    /**
+     * Timezone in which this provider has holidays defined
+     */
+    const TIMEZONE = 'Europe/Helsinki';
+
+    /**
+     * Locale that is considered common for this provider
+     */
+    const LOCALE = 'fi_FI';
+
+}